Veri Doğrulama(Data Validation) özelliği eklendi. -API versiyonlama yapılarak endpoint yapısı oluşturuldu.

+++ Controller/CommentController.cs
@@ -9,8 +9,9 @@ using TarifPaylasim.Mappers;
 
 namespace TarifPaylasim.Controller
 {
-    [Route("api/comment")]
-    [ApiController]
+   [ApiController]
+   [ApiVersion("1.0")]
+   [Route("api/v{version:apiVersion}/comment")]
     public class CommentController : ControllerBase
     {
         private readonly ICommentRepository _commentRepo;
@@ -25,15 +26,25 @@ namespace TarifPaylasim.Controller
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comments = await _commentRepo.GetAllAsync();
 
             var commentDto = comments.Select(x => x.toCommentDto());
             return Ok(comments);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if(ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepo.GetByIdAsync(id);
             if (comment == null)
             {
@@ -43,22 +54,33 @@ namespace TarifPaylasim.Controller
             return Ok(comment.toCommentDto());
         }
 
-        [HttpPost("{recipeId}")]
+        [HttpPost("{recipeId:int}")]
         public async Task<IActionResult> Create([FromRoute] int recipeId, [FromBody] CreateCommentDto commentDto)
         {
+            if(ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             if (!await _recipeRepo.RecipeExists(recipeId))
             {
                 return BadRequest("Tarif bilgisi bulunamadı");
             }
 
             var commentModel = commentDto.ToCommentFromCreateDTO(recipeId);
+            var comment = await _commentRepo.CreateAsync(commentModel);
             return CreatedAtAction(nameof(GetById), new { id = commentModel.Id }, commentModel.toCommentDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if(ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate(id));
 
             if (comment == null)
@@ -73,6 +95,11 @@ namespace TarifPaylasim.Controller
         [Route("{id}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if(ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
            var comment = await _commentRepo.DeleteAsync(id);
             if (comment == null)
             {
